refactor dns provider registration to centralized registry
